feat(card-basic): change basic card styles to theme friendly
