+++ src/components/CardBasic/dim.scss
@@ -1,0 +1,11 @@
+$card-corner-radius: 0.8rem;
+$card-padding-x: (
+  small: 0.9rem,
+  medium: 1.15rem,
+  large: 1.4rem,
+);
+$card-padding-y: (
+  small: 0.4rem,
+  medium: 0.6rem,
+  large: 0.8rem,
+);
