+++ src/views/pageTypewriter/components/Portrait/ActionBar/ActionBar.vue
@@ -2,7 +2,7 @@
 import { computed } from "vue";
 import { useKeyboardLayout } from "../../../hooks";
 import variables from "./variables.module.scss";
-import CardBasic from "@/components/CardBasic.vue";
+import CardBasic from "@/components/CardBasic/CardBasic.vue";
 import SvgIcon from "@/components/SvgIcon.vue";
 import ImgKeyboard from "../../../assets/keyboard.svg";
