+++ src/views/pageTypewriter/components/Portrait/InputArea/InputArea.vue
@@ -3,7 +3,7 @@ import { computed } from "vue";
 import { ConfigConvertDirectionEnum } from "@/shared";
 import { useConfig, useTextInput } from "../../../hooks";
 import variables from "./variables.module.scss";
-import CardBasic from "@/components/CardBasic.vue";
+import CardBasic from "@/components/CardBasic/CardBasic.vue";
 import TextAreaBasic from "@/components/TextAreaBasic.vue";
 import SvgIcon from "@/components/SvgIcon.vue";
 import ImgClose from "../../../assets/close.svg";
